pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "luanti-las-to-map"
version = "0.0.1"
authors = [
  { name="Richard Henwood", email="rjhenwood@yahoo.co.uk" },
]
description = "A package to convert LAS files to Luanti maps"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "zstandard",
    "laspy",
    "laszip",
    "numpy",
    "Pillow",
]


[project.urls]
"Homepage" = "https://github.com/richardhenwood/luanti-las-to-map"
"Bug Tracker" = "https://github.com/richardhenwood/luanti-las-to-map/issues"
